Struct rusoto_dynamodbstreams::DynamoDbStreamsClient

pub struct DynamoDbStreamsClient { /* fields omitted */ }

A client for the Amazon DynamoDB Streams API.

Methods

impl DynamoDbStreamsClient

pub fn new(region: Region) -> DynamoDbStreamsClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> DynamoDbStreamsClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,