Struct rusoto_dynamodbstreams::Record

pub struct Record {
    pub aws_region: Option<String>,
    pub dynamodb: Option<StreamRecord>,
    pub event_id: Option<String>,
    pub event_name: Option<String>,
    pub event_source: Option<String>,
    pub event_version: Option<String>,
    pub user_identity: Option<Identity>,
}

A description of a unique event within a stream.

Fields
aws_region: Option<String>

The region in which the GetRecords request was received.

dynamodb: Option<StreamRecord>

The main body of the stream record, containing all of the DynamoDB-specific fields.

event_id: Option<String>

A globally unique identifier for the event that was recorded in this stream record.

event_name: Option<String>

The type of data modification that was performed on the DynamoDB table:

INSERT - a new item was added to the table.

MODIFY - one or more of an existing item's attributes were modified.

REMOVE - the item was deleted from the table

event_source: Option<String>

The AWS service from which the stream record originated. For DynamoDB Streams, this is aws:dynamodb.

event_version: Option<String>

The version number of the stream record format. This number is updated whenever the structure of Record is modified.

Client applications must not assume that eventVersion will remain at a particular value, as this number is subject to change at any time. In general, eventVersion will only increase as the low-level DynamoDB Streams API evolves.

user_identity: Option<Identity>

Items that are deleted by the Time to Live process after expiration have the following fields:

Records[].userIdentity.type

"Service"

Records[].userIdentity.principalId

"dynamodb.amazonaws.com"