Struct rusoto_dynamodbstreams::StreamDescription

pub struct StreamDescription {
    pub creation_request_date_time: Option<f64>,
    pub key_schema: Option<Vec<KeySchemaElement>>,
    pub last_evaluated_shard_id: Option<String>,
    pub shards: Option<Vec<Shard>>,
    pub stream_arn: Option<String>,
    pub stream_label: Option<String>,
    pub stream_status: Option<String>,
    pub stream_view_type: Option<String>,
    pub table_name: Option<String>,
}

Represents all of the data describing a particular stream.

Fields
creation_request_date_time: Option<f64>

The date and time when the request to create this stream was issued.

key_schema: Option<Vec<KeySchemaElement>>

The key attribute(s) of the stream's DynamoDB table.

last_evaluated_shard_id: Option<String>

The shard ID of the item where the operation stopped, inclusive of the previous result set. Use this value to start a new operation, excluding this value in the new request.

If LastEvaluatedShardId is empty, then the "last page" of results has been processed and there is currently no more data to be retrieved.

If LastEvaluatedShardId is not empty, it does not necessarily mean that there is more data in the result set. The only way to know when you have reached the end of the result set is when LastEvaluatedShardId is empty.

shards: Option<Vec<Shard>>

The shards that comprise the stream.

stream_arn: Option<String>

The Amazon Resource Name (ARN) for the stream.

stream_label: Option<String>

A timestamp, in ISO 8601 format, for this stream.

Note that LatestStreamLabel is not a unique identifier for the stream, because it is possible that a stream from another table might have the same timestamp. However, the combination of the following three elements is guaranteed to be unique:

the AWS customer ID.

the table name

the StreamLabel

stream_status: Option<String>

Indicates the current status of the stream:

ENABLING - Streams is currently being enabled on the DynamoDB table.

ENABLED - the stream is enabled.

DISABLING - Streams is currently being disabled on the DynamoDB table.

DISABLED - the stream is disabled.

stream_view_type: Option<String>

Indicates the format of the records within this stream:

KEYS_ONLY - only the key attributes of items that were modified in the DynamoDB table.

NEW_IMAGE - entire items from the table, as they appeared after they were modified.

OLD_IMAGE - entire items from the table, as they appeared before they were modified.

NEW_AND_OLD_IMAGES - both the new and the old images of the items from the table.

table_name: Option<String>

The DynamoDB table with which the stream is associated.