Struct rusoto_dynamodbstreams::StreamRecord

pub struct StreamRecord {
    pub approximate_creation_date_time: Option<f64>,
    pub keys: Option<HashMap<String, AttributeValue>>,
    pub new_image: Option<HashMap<String, AttributeValue>>,
    pub old_image: Option<HashMap<String, AttributeValue>>,
    pub sequence_number: Option<String>,
    pub size_bytes: Option<i64>,
    pub stream_view_type: Option<String>,
}

A description of a single data modification that was performed on an item in a DynamoDB table.

Fields
approximate_creation_date_time: Option<f64>

The approximate date and time when the stream record was created, in UNIX epoch time format.

keys: Option<HashMap<String, AttributeValue>>

The primary key attribute(s) for the DynamoDB item that was modified.

new_image: Option<HashMap<String, AttributeValue>>

The item in the DynamoDB table as it appeared after it was modified.

old_image: Option<HashMap<String, AttributeValue>>

The item in the DynamoDB table as it appeared before it was modified.

sequence_number: Option<String>

The sequence number of the stream record.

size_bytes: Option<i64>

The size of the stream record, in bytes.

stream_view_type: Option<String>

The type of data from the modified DynamoDB item that was captured in this stream record:

KEYS_ONLY - only the key attributes of the modified item.

NEW_IMAGE - the entire item, as it appeared after it was modified.

OLD_IMAGE - the entire item, as it appeared before it was modified.

NEW_AND_OLD_IMAGES - both the new and the old item images of the item.